% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklegal.R
\name{checklegal}
\alias{checklegal}
\title{Takes an age and country and returns whether a person is legal to drink in that country}
\usage{
checklegal(age, country)
}
\arguments{
\item{country}{}
}
\value{
String of legal, not legal, or varies by conditions
}
\description{
Takes an age and country and returns whether a person is legal to drink in that country
}
\examples{
checklegal(16, "Canada")
}
