% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_legal.R
\name{countries}
\alias{countries}
\title{Find countries in the world where you can legally drink alc}
\usage{
countries(age)
}
\arguments{
\item{your_age}{}
}
\value{
The list of countries where you can legally drink alc
}
\description{
Find countries in the world where you can legally drink alc
}
\examples{
countries(15)

}
