% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploaddata.R
\docType{data}
\name{drinking}
\alias{drinking}
\title{Spreadsheet of Drinking Laws}
\format{A data frame with 179 rows and 5 columns
\describe{
  \item{index}{unique number for each country}
  \item{country}{name of country}
  \item{onpremisenum}{integer of minimum drinking age, -1 for varied laws, 0 for all-ages legal, 1000000 for illegal
  \item{variesby}{string of specific laws if they vary}
  \item{notes}{string of additional notes (not used)}}
  ...
}}
\description{
This data set contains all countries in the world and its drinking laws and minimum ages
}
