% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_legal.R
\name{variesby}
\alias{variesby}
\title{Find countries where legal drinking age varies by religion/jurisdiction/province/beverages}
\usage{
variesby(any)
}
\arguments{
\item{any}{}
}
\value{
A table of countries where legal drinking age varies by religion/jurisdiction/province/beverages
        Detail under NOTES column
}
\description{
Find countries where legal drinking age varies by religion/jurisdiction/province/beverages
}
\examples{
variesby(15)
}
